qt6: qttools: remove spurious clang dependency in dey-toolchain SDK
Typically, recipes that support clang builds only pull in the required clang
dependencies when the TOOLCHAIN variable is expliciltly set to "clang". The
qttools recipe in meta-qt6 differs from this by adding the dependency as soon
as meta-clang is present in the bblayers, regardless of the selected toolchain.

This causes all of the nativesdk clang packages to get included in our SDK
(along with the gcc ones), causing SDK bloat and spurious build errors on build
machines with limited resources.

Make sure that the dependency is only added when the clang TOOLCHAIN is
explicitly selected.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>