ccmp1/ccmp2: clear ALSA_LIST with override and move it to include file
ALSA_LIST variable is weakly set through an override (stm32mp1common or
stm32mp2common). When the override is in place, the hard assignment of
the plain variable is not enough, and the value weakly set in
meta-st-stm32mp layer takes precedence over the plain variable hard
assignment.
Clear the variable with the override and move it from machine config files
to the include file.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>