tf-a-stm32mp: use MACHINE variable for final metadata binary symlink
Currently, the final metadata symlink is composed using the TF-A Device Tree
configuration, which includes memory variant details. However, these variants
are not relevant for the metadata binary.

To avoid generating multiple redundant metadata files or using confusing names,
this commit updates the symlink to be composed using the MACHINE variable
instead.

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>