swupdate: fix regression in 2024.12.1
This version of swupdate has a bug that happens if the root of sw-description
is redirected via a link, which is the case is some of our sw-description
templates (such as the one we use for file updates). Backport a fix from
v2025.05.

https://onedigi.atlassian.net/browse/ADK4A-1957

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
(cherry picked from commit 5bdd59a6473257bbbe69f2b81fd70ba4fa26dcfd)