vsftpd: add support for SSL/TLS encrypted FTP sessions
This commit enables SSL/TLS support in vsftpd, allowing FTP communications to
be encrypted for improved security.

SSL/TLS support is enabled by default, but to preserve compatibility with
clients that do not support encrypted connections, this feature can be disabled
at runtime. Users can simply comment or uncomment a few lines in the
`vsftpd.conf` configuration file to toggle the behavior.

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>