weston-init: wait for a wayland socket before logging in
The login prompt appears before Wayland is fully initialized and
has created a wayland socket.

Logging in too early as root in this scenario caused the
WAYLAND_DISPLAY environment variable to be left empty. As a
consequence, gstreamer failed to use waylandsink to print contents
in the display.

Introduce a 10-seconds polling loop to wait for the wayland socket to
be available before proceeding with the login.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>