ccmp25-dvk: integrate Sony IMX335 MIPI camera in default device tree
As the OmniVision OV5640 camera is now deprecated and no longer supported by
most vendors, this commit moves its support to a separate Device Tree overlay,
allowing it to be used if needed.

Instead, the Sony IMX335 MIPI camera is integrated into the default device
tree as the default supported camera for the CCMP25-DVK platform.

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>