kernel-module-nxp-wlan: add watch-regdomain service
Add service to automatically detect changes in the global regulatory
domain and force a PHY regulatory domain change.

This allows detecting regulatory domain changes based on beacon
information when 802.11d is enabled and instructing the wireless
driver to download the rgpower firmware file corresponding to the
selected country.

If the selected country is not one of the supported ones, Worldwide
rgpower_WW.bin file will be downloaded by default.

Run the check service every 5 seconds through a systemd timer.

https://onedigi.atlassian.net/browse/DEL-8974

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>