linux-dey: tsn: fix tsn kernel support
fix missing option to include the required kernel options for tsn support.
That support was added in commit 37f5db42ac6d for kernel 6.1, but it
was missed for kernel 6.6.
Add the support to include the required fragment, regardless the kernel
version.
While on it, update the tsn config to match kernel 6.6.

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>