stm-st-stm32mp: m33projects-stm32mp2: Add custom sign-stm32mp bbclass to fix build issue
This commit imports the Digi custom version of sign-stm32mp bbclass to ensure
that the search_path() function does not raise a build exception if the signing
tool or keys are not present in the PATH before starting the build process.

In our case, we do not need to manually install the tools or generate the keys
beforehand, as this is automatically handled by Yocto in our DEY distribution.

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>