mkproject: ignore meta-digi-arm dir when other layer is selected
Default value of 'MKP_CONFIGPATH' points to a config path in the
meta-digi/meta-digi-arm layer.

When mkproject.sh is called with the '-m' argument, another layer is
provided to build the project from. In this layer the conf/templates
new structure must be respected, but it is likely it does not contain a
meta-digi-arm directory on top.

Keep the 'meta-digi-arm' directory in the config path by default, but
remove it when a different layer is provided through an argument.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>