u-boot-dey: adjust KCONFIG_CONFIG_ROOTDIR to be able to run do_menuconfig
In Yocto 4.0, the do_menuconfig task for u-boot recipes is broken: the shell
window that's supposed to contain the menuconfig menu appears for a fraction
of a second and disappears immediately after, making the feature unusable. This
is because KCONFIG_CONFIG_ROOTDIR, the variable containing the path where
"make menuconfig" gets run, is pointing to the wrong directory.

Said variable defaults to ${B} (the package's "build" directory), which is
correct in most cases, but since u-boot packages can have multiple
configurations (each with its own build), each of these configurations has
its own subdirectory inside of ${B}. This is true even for platforms with only
one u-boot configuration. The result is that "make menuconfig" is run in a
directory without a Makefile, and the shell exits immediately after.

Fix this by partially porting logic from poky commit 0241bdb49c52, in which
KCONFIG_CONFIG_ROOTDIR is adjusted for u-boot packages so that do_menuconfig
works properly in platforms with a single u-boot configuration.

In the case of platforms with multiple configurations (ccimx6, ccimx6qp and
ccimx6ul), do_menuconfig cannot be used. In Yocto 5.0, attempting to run the
task for these platforms will trigger an error. Mimic this behavior by
triggering a similar error on platforms with multiple configurations.

https://onedigi.atlassian.net/browse/DUB-1109

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>