wpewebkit: port custom Digi patches to 2.46.X
Out of the three patches we currently apply to v2.44.4, two are no longer
needed for v2.46.7:

    * 0001-Build-fix-when-LAYER_BASED_SVG_ENGINE-is-off.patch is already
      included in v2.46.X releases
    * 0001-DMABufVideoSinkGStreamer-disable-sink-unconditionall.patch is no
      longer needed, since the GL sink is chosen by default in v2.46.7

Only port 0001-UIProcess-WebProcessPool-always-swap-process-when-us.patch to
avoid performance issues on platforms with memory restrictions.

Keep the v2.44.4 patches in case customers want to use said version, and
clearly separate the patchsets for both versions.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>