webkit: revert wpewebkit downgrade to v2.44.4
This partially reverts commit 6a7e9fa9e40a. We initially downgraded to v2.44.4
to fix an issue with mouse clicks in our ConnectCore demo, but these issues
have been fixed after cog was upgraded to v0.18.5 in meta-webkit.

Remove the downgrade to use the newest v2.46.X versions of wpewebkit in the
scarthgap branch of meta-webkit, currently at v2.46.7.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>