connectcore-demo-example: remove browser dependencies from browser-based pkgs
Commit 87b73f3f5dba consolidated all connectcore-demo-example recipes into a
single recipe with multiple packages. Even though we choose which of these
packages to install depending on the image type, all packages (along with their
dependencies) get built regardless of which package ends up getting installed.
This means that the browsers (webkit/chromium) will get built as long as their
respective layer is in the bblayers, regardless of the image type.

For example:
    * wpewebkit will get built for all platforms with webkit support,
      regardless of the image (qt, flutter, core-image-base...)
    * chromium-ozone-wayland will get built for the ccimx95-dvk, even for qt
      and flutter images

Since we already include the main webkit and chromium packages in their
respective image's packagegroup, simply remove the dependencies from the
connectcore-demo-example recipe to avoid build overhead and unexpected errors.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>