cccs: don't change default client_cert_path for the ccimx6ul
Commit 619ca8b2a8d6 changed the ccimx6ul's default endpoint URL and client cert
path in order to prevent performance issues if the client cert is erased before
registering the device.

In reality, the client cert path doesn't really matter after this change since
the remotemanager.digi.com endpoint won't make use of it, and we now ensure
that /mnt/data is writable; so we might as well keep using that path.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>